requirements.txt
fastapi
uvicorn
langchain-google-genai
pydantic
python-dotenv
streamlit
langchain-pinecone
pinecone-client
pypdf
requests
langchain-community
sentence-transformers
langchain-huggingface